# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  crawling: 'crawling',
  paused: 'paused',
  error: 'error',
} as const;

export interface Project {
  id: string;
  name: string;
  url: string;
  keywords?: string[];
  /** @nullable */
  country?: string | null;
  /** @nullable */
  language?: string | null;
  status: ProjectStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  /** @nullable */
  lastCrawlAt: string | null;
  pagesCrawled: number;
  issueCount: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  url: string;
  keywords?: string[];
  country?: string;
  language?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  keywords?: string[];
  country?: string;
  language?: string;
  status?: ProjectUpdateStatus;
}

export interface CrawlInput {
  /**
     * @minimum 1
     * @maximum 500
     */
  maxPages?: number;
}

export type CrawlStatus = typeof CrawlStatus[keyof typeof CrawlStatus];


export const CrawlStatus = {
  queued: 'queued',
  crawling: 'crawling',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Crawl {
  id: string;
  projectId: string;
  status: CrawlStatus;
  pagesDiscovered: number;
  pagesAnalyzed: number;
  issuesFound: number;
  /** @nullable */
  score?: number | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface ScorePoint {
  date: string;
  score: number;
}

export interface CategoryScore {
  category: string;
  score: number;
}

export interface CountPoint {
  label: string;
  count: number;
}

export interface DashboardMetrics {
  pages: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
  brokenLinks: number;
  missingMetadata: number;
  duplicateContent: number;
}

export interface Dashboard {
  project: Project;
  scoreHistory: ScorePoint[];
  categoryScores: CategoryScore[];
  severityCounts: CountPoint[];
  metrics: DashboardMetrics;
}

export interface Page {
  id: string;
  url: string;
  status: number;
  score: number;
  /** @nullable */
  title: string | null;
  /** @nullable */
  h1: string | null;
  /** @nullable */
  metaDescription: string | null;
  issueCount: number;
}

export interface PageList {
  items: Page[];
  total: number;
  page: number;
  limit: number;
}

export type PageReportHeadings = {
  h1?: string[];
  h2?: string[];
  h3?: string[];
};

export type PageReportImages = {
  total?: number;
  missingAlt?: number;
};

export type PageReportLinks = {
  internal?: number;
  external?: number;
  broken?: number;
};

export type PageReportSocial = {
  openGraph?: boolean;
  twitterCard?: boolean;
};

export type PageReportStructuredData = {
  present?: boolean;
  types?: string[];
};

export type SEOIssueSeverity = typeof SEOIssueSeverity[keyof typeof SEOIssueSeverity];


export const SEOIssueSeverity = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface SEOIssue {
  id: string;
  severity: SEOIssueSeverity;
  category: string;
  title: string;
  url: string;
  description: string;
  whyItMatters: string;
  recommendedFix: string;
  reviewed: boolean;
  /** @nullable */
  recommendationId?: string | null;
}

export type PageReport = Page & {
  headings: PageReportHeadings;
  images: PageReportImages;
  links: PageReportLinks;
  social: PageReportSocial;
  structuredData: PageReportStructuredData;
  issues: SEOIssue[];
};

export interface IssueUpdate {
  reviewed: boolean;
}

export interface AIRecommendation {
  id: string;
  issueId: string;
  problem: string;
  whyItMatters: string;
  recommendedSolution: string;
  suggestedImplementation: string;
  createdAt: string;
}

export interface AssistantInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  question: string;
}

export type AssistantResponseSourcesItem = {
  label: string;
  url: string;
};

export interface AssistantResponse {
  answer: string;
  sources: AssistantResponseSourcesItem[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;