# lib/api-zod/src/generated/types/aIRecommendation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface AIRecommendation {
  id: string;
  issueId: string;
  problem: string;
  whyItMatters: string;
  recommendedSolution: string;
  suggestedImplementation: string;
  createdAt: Date;
}