# lib/api-zod/src/generated/types/assistantResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantResponseSourcesItem } from './assistantResponseSourcesItem';

export interface AssistantResponse {
  answer: string;
  sources: AssistantResponseSourcesItem[];
}