# lib/api-zod/src/generated/types/categoryScore.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryScore {
  category: string;
  score: number;
}