# lib/api-zod/src/generated/types/crawl.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { CrawlStatus } from './crawlStatus';

export interface Crawl {
  id: string;
  projectId: string;
  status: CrawlStatus;
  pagesDiscovered: number;
  pagesAnalyzed: number;
  issuesFound: number;
  /** @nullable */
  score?: number | null;
  startedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  error?: string | null;
}