# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryScore } from './categoryScore';
import type { CountPoint } from './countPoint';
import type { DashboardMetrics } from './dashboardMetrics';
import type { Project } from './project';
import type { ScorePoint } from './scorePoint';

export interface Dashboard {
  project: Project;
  scoreHistory: ScorePoint[];
  categoryScores: CategoryScore[];
  severityCounts: CountPoint[];
  metrics: DashboardMetrics;
}