# lib/api-zod/src/generated/types/pageList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { Page } from './page';

export interface PageList {
  items: Page[];
  total: number;
  page: number;
  limit: number;
}