# lib/api-zod/src/generated/types/pageReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { Page } from './page';
import type { PageReportHeadings } from './pageReportHeadings';
import type { PageReportImages } from './pageReportImages';
import type { PageReportLinks } from './pageReportLinks';
import type { PageReportSocial } from './pageReportSocial';
import type { PageReportStructuredData } from './pageReportStructuredData';
import type { SEOIssue } from './sEOIssue';

export type PageReport = Page & {
  headings: PageReportHeadings;
  images: PageReportImages;
  links: PageReportLinks;
  social: PageReportSocial;
  structuredData: PageReportStructuredData;
  issues: SEOIssue[];
};