# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEOlytic AI SEO Auditor API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectUpdateStatus } from './projectUpdateStatus';

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  keywords?: string[];
  country?: string;
  language?: string;
  status?: ProjectUpdateStatus;
}